+++ src/main/kotlin/kscript/app/appdir/Cache.kt
@@ -7,22 +7,37 @@ import kscript.app.util.*
 import org.apache.commons.codec.digest.DigestUtils
 import java.net.URI
 
-class Cache(private val path: OsPath) {
+class Cache(private val cacheBasePath: OsPath) {
     fun getOrCreateIdeaProject(digest: String, creator: (OsPath) -> OsPath): OsPath {
-        return directoryCache(path.resolve("idea_$digest"), creator)
+        val path = cacheBasePath.resolve("idea_$digest")
+
+        return if (path.exists()) {
+            path
+        } else {
+            path.createDirectories()
+            creator(path)
+        }
     }
 
-    fun getOrCreatePackage(digest: String, creator: (OsPath) -> OsPath): OsPath {
-        return directoryCache(path.resolve("package_$digest"), creator)
+    fun getOrCreatePackage(digest: String, scriptName: String, creator: (OsPath, OsPath) -> OsPath): OsPath {
+        val path = cacheBasePath.resolve("package_$digest")
+        val cachedPackageFile = path.resolve("build/libs/$scriptName")
+
+        return if (cachedPackageFile.exists()) {
+            cachedPackageFile
+        } else {
+            path.createDirectories()
+            creator(path, cachedPackageFile)
+        }
     }
 
     fun getOrCreateJar(digest: String, creator: (OsPath) -> JarArtifact): JarArtifact {
-        val directory = path.resolve("jar_$digest")
+        val directory = cacheBasePath.resolve("jar_$digest")
         val cachedJarArtifact = directory.resolve("jarArtifact.descriptor")
 
         return if (cachedJarArtifact.exists()) {
             val jarArtifactLines = cachedJarArtifact.readText().lines()
-            JarArtifact(OsPath.createOrThrow(path.nativeType, jarArtifactLines[0]), jarArtifactLines[1])
+            JarArtifact(OsPath.createOrThrow(cacheBasePath.nativeType, jarArtifactLines[0]), jarArtifactLines[1])
         } else {
             directory.createDirectories()
             val jarArtifact = creator(directory)
@@ -34,7 +49,7 @@ class Cache(private val path: OsPath) {
     fun getOrCreateUriItem(uri: URI, creator: (URI, OsPath) -> Content): Content {
         val digest = DigestUtils.md5Hex(uri.toString())
 
-        val directory = path.resolve("uri_$digest")
+        val directory = cacheBasePath.resolve("uri_$digest")
         val descriptorFile = directory.resolve("uri.descriptor")
         val contentFile = directory.resolve("uri.content")
 
@@ -61,7 +76,7 @@ class Cache(private val path: OsPath) {
     }
 
     fun getOrCreateDependencies(digest: String, creator: () -> Set<OsPath>): Set<OsPath> {
-        val directory = path.resolve("dependencies_$digest")
+        val directory = cacheBasePath.resolve("dependencies_$digest")
         val contentFile = directory.resolve("dependencies.content")
 
         if (directory.exists()) {
@@ -69,7 +84,7 @@ class Cache(private val path: OsPath) {
                 contentFile.readText()
                     .lines()
                     .filter { it.isNotEmpty() }
-                    .map { OsPath.createOrThrow(path.nativeType, it) }
+                    .map { OsPath.createOrThrow(cacheBasePath.nativeType, it) }
                     .toSet()
 
             //Recheck cached paths - if there are missing artifacts skip the cached values
@@ -83,14 +98,5 @@ class Cache(private val path: OsPath) {
         contentFile.writeText(dependencies.joinToString("\n") { it.toString() })
 
         return dependencies
-    }
-
-    private fun directoryCache(path: OsPath, creator: (OsPath) -> OsPath): OsPath {
-        return if (path.exists()) {
-            path
-        } else {
-            path.createDirectories()
-            creator(path)
-        }
     }
 }