+++ src/main/kotlin/kscript/app/code/GradleTemplates.kt
@@ -51,60 +51,60 @@ object GradleTemplates {
         val baseName = script.scriptName
 
         return """
-        import java.io.*
-        import java.lang.System
-        import java.nio.file.Files
-        import java.nio.file.Paths
-
-        plugins {
-            id("org.jetbrains.kotlin.jvm") version "$kotlinVersion"
-            application
-        }
-
-        repositories {
-            mavenLocal()
-            mavenCentral()
-        ${createGradleRepositoriesSection(script.repositories).prependIndent()}
-        }
-
-        tasks.jar {
-            manifest {
-                attributes["Main-Class"] = "$capsuleApp"
-            }
-            baseName = "$baseName"
-            configurations["compileClasspath"].forEach { file: File ->
-                from(zipTree(file.absoluteFile))
-            }
-            duplicatesStrategy = DuplicatesStrategy.INCLUDE
-        }
-
-        tasks.register("makeScript") {
-            dependsOn(":jar")
-            doLast {
-                val headerDir = layout.projectDirectory.toString()
-                val jarFileName = layout.buildDirectory.file("libs/$baseName.jar").get().toString()
-                val outFileName = layout.buildDirectory.file("libs/$baseName").get().toString()
-                val lineSeparator = System.getProperty("line.separator").encodeToByteArray()
-                val headerPath = Paths.get(headerDir).resolve("exec_header.sh")
-                val headerBytes = Files.readAllBytes(headerPath)
-                val jarBytes = Files.readAllBytes(Paths.get(jarFileName))
-                val outFile = Paths.get(outFileName).toFile()
-                val fileStream = FileOutputStream(outFile)
-
-                fileStream.write(headerBytes)
-                fileStream.write(lineSeparator)
-                fileStream.write(jarBytes)
-                fileStream.close()
-            }
-        }
-
-        dependencies {
-            implementation(files("${jarArtifact.path.parent.resolve("scriplet.jar")}"))
-        ${createGradleDependenciesSection(extendedDependencies).prependIndent()}
-        }
-
-        $kotlinOptions
-        """.trimStart('\n').trimIndent()
+            |import java.io.*
+            |import java.lang.System
+            |import java.nio.file.Files
+            |import java.nio.file.Paths
+            |
+            |plugins {
+            |    id("org.jetbrains.kotlin.jvm") version "$kotlinVersion"
+            |    application
+            |}
+            |
+            |repositories {
+            |    mavenLocal()
+            |    mavenCentral()
+            |${createGradleRepositoriesSection(script.repositories).prependIndent()}
+            |}
+            |
+            |tasks.jar {
+            |    manifest {
+            |        attributes["Main-Class"] = "$capsuleApp"
+            |    }
+            |    baseName = "$baseName"
+            |    configurations["compileClasspath"].forEach { file: File ->
+            |        from(zipTree(file.absoluteFile))
+            |    }
+            |    duplicatesStrategy = DuplicatesStrategy.INCLUDE
+            |}
+            |
+            |tasks.register("makeScript") {
+            |    dependsOn(":jar")
+            |    doLast {
+            |        val headerDir = layout.projectDirectory.toString()
+            |        val jarFileName = layout.buildDirectory.file("libs/$baseName.jar").get().toString()
+            |        val outFileName = layout.buildDirectory.file("libs/$baseName").get().toString()
+            |        val lineSeparator = System.getProperty("line.separator").encodeToByteArray()
+            |        val headerPath = Paths.get(headerDir).resolve("exec_header.sh")
+            |        val headerBytes = Files.readAllBytes(headerPath)
+            |        val jarBytes = Files.readAllBytes(Paths.get(jarFileName))
+            |        val outFile = Paths.get(outFileName).toFile()
+            |        val fileStream = FileOutputStream(outFile)
+            |
+            |        fileStream.write(headerBytes)
+            |        fileStream.write(lineSeparator)
+            |        fileStream.write(jarBytes)
+            |        fileStream.close()
+            |    }
+            |}
+            |
+            |dependencies {
+            |    implementation(files("${jarArtifact.path.parent.resolve("scriplet.jar")}"))
+            |${createGradleDependenciesSection(extendedDependencies).prependIndent()}
+            |}
+            |
+            |$kotlinOptions
+            """.trimStart().trimMargin()
     }
 
     private fun createGradleRepositoryCredentials(repository: Repository): String {